impl StepEstimator
