+++ RefraSin.TEPSolver/EquationSystem/LagrangianGradient.cs
@@ -159,53 +159,4 @@ public static class LagrangianGradient
         );
     }
 
-    public static StepVector GuessSolution(SolutionState currentState) =>
-        new(YieldInitialGuess(currentState).ToArray(),
-            new StepVectorMap(currentState));
-
-    private static IEnumerable<double> YieldInitialGuess(SolutionState currentState) =>
-        YieldNodeUnknownsInitialGuess(currentState)
-            .Concat(
-                YieldContactNodeUnknownsInitialGuess(currentState)
-            )
-            .Concat(
-                YieldContactUnknownsInitialGuess(currentState)
-            ).Concat(
-                YieldGlobalUnknownsInitialGuess()
-            );
-
-    private static IEnumerable<double> YieldGlobalUnknownsInitialGuess()
-    {
-        yield return 1;
-    }
-
-    private static IEnumerable<double> YieldContactUnknownsInitialGuess(SolutionState currentState)
-    {
-        foreach (var _ in currentState.Contacts)
-        {
-            yield return 0;
-            yield return 0;
-            yield return 0;
-        }
-    }
-
-    private static IEnumerable<double> YieldNodeUnknownsInitialGuess(SolutionState currentState)
-    {
-        foreach (var node in currentState.Nodes)
-        {
-            yield return node.GuessNormalDisplacement();
-            yield return node.GuessFluxToUpper();
-            yield return 1;
-        }
-    }
-
-    private static IEnumerable<double> YieldContactNodeUnknownsInitialGuess(SolutionState currentState)
-    {
-        foreach (var _ in currentState.Nodes.OfType<ContactNodeBase>())
-        {
-            yield return 0;
-            yield return 1;
-            yield return 1;
-        }
-    }
 }
