+++ RefraSin.TEPSolver/ParticleModel/NodeBase.cs
@@ -156,33 +156,6 @@ public abstract class NodeBase : INode, INodeGeometry, INodeGradients, INodeMate
 
     private NormalTangential<double>? _volumeGradient;
 
-    public double GuessFluxToUpper()
-    {
-        var x1 = -Lower.Coordinates.R * Sin(AngleDistance.ToLower);
-        var x3 = Upper.Coordinates.R * Sin(AngleDistance.ToUpper);
-        var y1 = Lower.Coordinates.R * Cos(AngleDistance.ToLower);
-        var y2 = Coordinates.R;
-        var y3 = Upper.Coordinates.R * Cos(AngleDistance.ToUpper);
-
-        var curvature = -(x3 * y1 + x1 * y2 - x3 * y2 - x1 * y3) / (Pow(x1, 2) * x3 - x1 * Pow(x3, 2));
-        var curvatureGibbs = GibbsEnergyGradient.Normal / (SurfaceDistance.ToUpper + SurfaceDistance.ToLower) / SurfaceEnergy.ToUpper;
-
-        var vacancyConcentrationGradient = -Particle.Material.EquilibriumVacancyConcentration
-                                         / (SolverSession.GasConstant * SolverSession.Temperature)
-                                         * (Upper.GibbsEnergyGradient.Normal - GibbsEnergyGradient.Normal)
-                                         * Particle.Material.MolarVolume
-                                         / Pow(SurfaceDistance.ToUpper, 2);
-        return -SurfaceDiffusionCoefficient.ToUpper * vacancyConcentrationGradient;
-    }
-
-    public double GuessNormalDisplacement()
-    {
-        var fluxBalance = GuessFluxToUpper() - Lower.GuessFluxToUpper();
-
-        var displacement = 2 * fluxBalance / ((SurfaceDistance.ToUpper + SurfaceDistance.ToLower) * Sin(SurfaceVectorAngle.Normal));
-        return displacement;
-    }
-
     public abstract NodeBase ApplyTimeStep(StepVector stepVector, double timeStepWidth, Particle particle);
 
     public override string ToString() =>
