+++ RefraSin.TEPSolver/StepEstimators/IStepEstimator.cs
@@ -1,7 +1,7 @@
 using RefraSin.ProcessModel;
 using RefraSin.TEPSolver.StepVectors;
 
-namespace RefraSin.TEPSolver.EquationSystem;
+namespace RefraSin.TEPSolver.StepEstimators;
 
 public interface IStepEstimator
 {
